builder to add vimeo videos with ids of any length
https://github.com/Esri/map-tour-storytelling-template-js/issues/26

When adding a new map tour point in the builder and selecting a vimeo
video, the builder only worked correctly in finding a thumbnail and
video if the id was exactly 8 digits. This isn't always the case with
vimeo videos. The latest videos now have 9 digits and there are even
videos in the single digits. Made a single change in two places to allow
for arbitrary length ids.